requirements.txt
click
tensorflow-gpu==2.0.1
tensorflow-addons==0.5
scikit-image=0.15
numpy
matplotlib
tqdm